reloadable

+++ source/common/router/upstream_request.cc
@@ -144,6 +144,11 @@ void CodecFilter::CodecBridge::decode1xxHeaders(Http::ResponseHeaderMapPtr&& hea
 // This is the response headers arriving from the codec. Send them through the filter manager.
 void CodecFilter::CodecBridge::decodeHeaders(Http::ResponseHeaderMapPtr&& headers,
                                              bool end_stream) {
+  // TODO(rodaine): This is actually measuring after the headers are parsed and not the first
+  // byte.
+  filter_.request_.upstreamTiming().onFirstUpstreamRxByteReceived(
+      filter_.request_.parent_.callbacks()->dispatcher().timeSource());
+
   filter_.request_.maybeEndDecode(end_stream);
   filter_.callbacks_->encodeHeaders(std::move(headers), end_stream,
                                     StreamInfo::ResponseCodeDetails::get().ViaUpstream);
@@ -218,6 +223,8 @@ UpstreamRequest::UpstreamRequest(RouterFilterInterface& parent,
       reset_stream_(false),
       record_timeout_budget_(parent_.cluster()->timeoutBudgetStats().has_value()),
       cleaned_up_(false), had_upstream_(false),
+      allow_upstream_filters_(
+          Runtime::runtimeFeatureEnabled("envoy.reloadable_features.allow_upstream_filters")),
       stream_options_({can_send_early_data, can_use_http3}) {
   if (parent_.config().start_child_span_) {
     span_ = parent_.callbacks()->activeSpan().spawnChild(
@@ -237,6 +244,11 @@ UpstreamRequest::UpstreamRequest(RouterFilterInterface& parent,
   if (cluster_info.has_value()) {
     stream_info_.setUpstreamClusterInfo(*cluster_info);
   }
+  if (!allow_upstream_filters_) {
+    return;
+  }
+
+  // Set up the upstream filter manager.
   fm_callbacks_ = std::make_unique<UpstreamRequestFilterManagerCallbacks>(*this);
   filter_manager_ = std::make_unique<UpstreamFilterManager>(
       *fm_callbacks_, parent_.callbacks()->dispatcher(), *parent_.callbacks()->connection(),
@@ -263,7 +275,9 @@ void UpstreamRequest::cleanUp() {
   }
   cleaned_up_ = true;
 
-  filter_manager_->destroyFilters();
+  if (allow_upstream_filters_) {
+    filter_manager_->destroyFilters();
+  }
 
   if (span_ != nullptr) {
     Tracing::HttpTracerUtility::finalizeUpstreamSpan(*span_, upstream_headers_.get(),
@@ -321,7 +335,9 @@ void UpstreamRequest::cleanUp() {
     parent_.cluster()->stats().upstream_flow_control_drained_total_.inc();
     --downstream_data_disabled_;
   }
-  parent_.callbacks()->dispatcher().deferredDelete(std::move(fm_callbacks_));
+  if (allow_upstream_filters_) {
+    parent_.callbacks()->dispatcher().deferredDelete(std::move(fm_callbacks_));
+  }
 }
 
 // This is called by the FilterManager when all filters have processed 1xx headers. Forward them
@@ -341,9 +357,6 @@ void UpstreamRequest::decodeHeaders(Http::ResponseHeaderMapPtr&& headers, bool e
   ENVOY_STREAM_LOG(trace, "upstream response headers:\n{}", *parent_.callbacks(), *headers);
   ScopeTrackerScopeState scope(&parent_.callbacks()->scope(), parent_.callbacks()->dispatcher());
 
-  // TODO(rodaine): This is actually measuring after the headers are parsed and not the first
-  // byte.
-  upstreamTiming().onFirstUpstreamRxByteReceived(parent_.callbacks()->dispatcher().timeSource());
   resetPerTryIdleTimer();
 
   addResponseHeadersSize(headers->byteSize());
@@ -365,6 +378,13 @@ void UpstreamRequest::decodeHeaders(Http::ResponseHeaderMapPtr&& headers, bool e
     return;
   }
 
+  if (!allow_upstream_filters_) {
+    // TODO(rodaine): This is actually measuring after the headers are parsed and not the first
+    // byte.
+    upstreamTiming().onFirstUpstreamRxByteReceived(parent_.callbacks()->dispatcher().timeSource());
+    maybeEndDecode(end_stream);
+  }
+
   awaiting_headers_ = false;
   if (!parent_.config().upstream_logs_.empty()) {
     upstream_headers_ = Http::createHeaderMap<Http::ResponseHeaderMapImpl>(*headers);
@@ -372,7 +392,7 @@ void UpstreamRequest::decodeHeaders(Http::ResponseHeaderMapPtr&& headers, bool e
   stream_info_.response_code_ = static_cast<uint32_t>(response_code);
 
   if (paused_for_connect_ && response_code == 200) {
-    codec_filter_->shipHeadersIfPaused(*parent_.downstreamHeaders());
+    encodeBodyAndTrailers();
     paused_for_connect_ = false;
   }
 
@@ -384,6 +404,9 @@ void UpstreamRequest::decodeData(Buffer::Instance& data, bool end_stream) {
   ScopeTrackerScopeState scope(&parent_.callbacks()->scope(), parent_.callbacks()->dispatcher());
 
   resetPerTryIdleTimer();
+  if (!allow_upstream_filters_) {
+    maybeEndDecode(end_stream);
+  }
   stream_info_.addBytesReceived(data.length());
   parent_.onUpstreamData(data, *this, end_stream);
 }
@@ -392,6 +415,9 @@ void UpstreamRequest::decodeTrailers(Http::ResponseTrailerMapPtr&& trailers) {
   ENVOY_STREAM_LOG(trace, "upstream response trailers:\n{}", *parent_.callbacks(), *trailers);
   ScopeTrackerScopeState scope(&parent_.callbacks()->scope(), parent_.callbacks()->dispatcher());
 
+  if (!allow_upstream_filters_) {
+    maybeEndDecode(true);
+  }
   if (!parent_.config().upstream_logs_.empty()) {
     upstream_trailers_ = Http::createHeaderMap<Http::ResponseTrailerMapImpl>(*trailers);
   }
@@ -430,11 +456,14 @@ void UpstreamRequest::onUpstreamHostSelected(Upstream::HostDescriptionConstShare
   parent_.onUpstreamHostSelected(host);
 }
 
-void UpstreamRequest::encodeHeaders(bool end_stream) {
+void UpstreamRequest::acceptHeadersFromRouter(bool end_stream) {
   ASSERT(!router_sent_end_stream_);
   router_sent_end_stream_ = end_stream;
 
   conn_pool_->newStream(this);
+  if (!allow_upstream_filters_) {
+    return;
+  }
 
   auto* headers = parent_.downstreamHeaders();
 
@@ -465,14 +494,63 @@ void UpstreamRequest::encodeHeaders(bool end_stream) {
   filter_manager_->decodeHeaders(*parent_.downstreamHeaders(), end_stream);
 }
 
-void UpstreamRequest::encodeData(Buffer::Instance& data, bool end_stream) {
+void UpstreamRequest::acceptDataFromRouterOld(Buffer::Instance& data, bool end_stream) {
+  if (!upstream_ || paused_for_connect_) {
+    ENVOY_STREAM_LOG(trace, "buffering {} bytes", *parent_.callbacks(), data.length());
+    if (!buffered_request_body_) {
+      buffered_request_body_ = parent_.callbacks()->dispatcher().getWatermarkFactory().createBuffer(
+          [this]() -> void { this->enableDataFromDownstreamForFlowControl(); },
+          [this]() -> void { this->disableDataFromDownstreamForFlowControl(); },
+          []() -> void { /* TODO(adisuissa): Handle overflow watermark */ });
+      buffered_request_body_->setWatermarks(parent_.callbacks()->decoderBufferLimit());
+    }
+
+    buffered_request_body_->move(data);
+  } else {
+    ASSERT(downstream_metadata_map_vector_.empty());
+
+    ENVOY_STREAM_LOG(trace, "proxying {} bytes", *parent_.callbacks(), data.length());
+    stream_info_.addBytesSent(data.length());
+    upstream_->encodeData(data, end_stream);
+    if (end_stream) {
+      upstreamTiming().onLastUpstreamTxByteSent(parent_.callbacks()->dispatcher().timeSource());
+    }
+  }
+}
+
+void UpstreamRequest::acceptDataFromRouter(Buffer::Instance& data, bool end_stream) {
   ASSERT(!router_sent_end_stream_);
   router_sent_end_stream_ = end_stream;
+
+  if (!allow_upstream_filters_) {
+    acceptDataFromRouterOld(data, end_stream);
+    return;
+  }
 
   filter_manager_->decodeData(data, end_stream);
 }
 
-void UpstreamRequest::encodeTrailers(Http::RequestTrailerMap& trailers) {
+void UpstreamRequest::acceptTrailersFromRouterOld(Http::RequestTrailerMap& trailers) {
+  router_sent_end_stream_ = true;
+  encode_trailers_ = true;
+
+  if (!upstream_) {
+    ENVOY_STREAM_LOG(trace, "buffering trailers", *parent_.callbacks());
+  } else {
+    ASSERT(downstream_metadata_map_vector_.empty());
+
+    ENVOY_STREAM_LOG(trace, "proxying trailers", *parent_.callbacks());
+    upstream_->encodeTrailers(trailers);
+    upstreamTiming().onLastUpstreamTxByteSent(parent_.callbacks()->dispatcher().timeSource());
+  }
+}
+
+void UpstreamRequest::acceptTrailersFromRouter(Http::RequestTrailerMap& trailers) {
+  if (!allow_upstream_filters_) {
+    acceptTrailersFromRouterOld(trailers);
+    return;
+  }
+
   ASSERT(!router_sent_end_stream_);
   router_sent_end_stream_ = true;
   encode_trailers_ = true;
@@ -480,7 +558,25 @@ void UpstreamRequest::encodeTrailers(Http::RequestTrailerMap& trailers) {
   filter_manager_->decodeTrailers(trailers);
 }
 
-void UpstreamRequest::encodeMetadata(Http::MetadataMapPtr&& metadata_map_ptr) {
+void UpstreamRequest::acceptMetadataFromRouterOld(Http::MetadataMapPtr&& metadata_map_ptr) {
+  if (!upstream_) {
+    ENVOY_STREAM_LOG(trace, "upstream_ not ready. Store metadata_map to encode later: {}",
+                     *parent_.callbacks(), *metadata_map_ptr);
+    downstream_metadata_map_vector_.emplace_back(std::move(metadata_map_ptr));
+  } else {
+    ENVOY_STREAM_LOG(trace, "Encode metadata: {}", *parent_.callbacks(), *metadata_map_ptr);
+    Http::MetadataMapVector metadata_map_vector;
+    metadata_map_vector.emplace_back(std::move(metadata_map_ptr));
+    upstream_->encodeMetadata(metadata_map_vector);
+  }
+}
+
+void UpstreamRequest::acceptMetadataFromRouter(Http::MetadataMapPtr&& metadata_map_ptr) {
+  if (!allow_upstream_filters_) {
+    acceptMetadataFromRouterOld(std::move(metadata_map_ptr));
+    return;
+  }
+
   filter_manager_->decodeMetadata(*metadata_map_ptr);
 }
 
@@ -626,7 +722,8 @@ void UpstreamRequest::onPoolReady(
 
   if (protocol) {
     stream_info_.protocol(protocol.value());
-  } else {
+  } else if (allow_upstream_filters_) {
+    // We only pause for CONNECT for HTTP upstreams. If this is a TCP upstream, unpause.
     paused_for_connect_ = false;
   }
 
@@ -676,8 +773,6 @@ void UpstreamRequest::onPoolReady(
   // downstream buffers are overrun. This may result in immediate watermark callbacks referencing
   // the encoder.
   parent_.callbacks()->addDownstreamWatermarkCallbacks(downstream_watermark_manager_);
-  // Make sure the local filter manager will do the same.
-  codec_filter_->callbacks_->addDownstreamWatermarkCallbacks(downstream_watermark_manager_);
 
   absl::optional<std::chrono::milliseconds> max_stream_duration;
   if (parent_.dynamicMaxStreamDuration().has_value()) {
@@ -686,17 +781,109 @@ void UpstreamRequest::onPoolReady(
     max_stream_duration = std::chrono::milliseconds(DurationUtil::durationToMilliseconds(
         upstream_host_->cluster().commonHttpProtocolOptions().max_stream_duration()));
   }
-
   if (max_stream_duration.has_value() && max_stream_duration->count()) {
     max_stream_duration_timer_ = parent_.callbacks()->dispatcher().createTimer(
         [this]() -> void { onStreamMaxDurationReached(); });
     max_stream_duration_timer_->enableTimer(*max_stream_duration);
   }
 
-  codec_filter_->shipHeadersIfPaused(*parent_.downstreamHeaders());
+  if (allow_upstream_filters_) {
+    // Make sure the local filter manager will do the same.
+    codec_filter_->callbacks_->addDownstreamWatermarkCallbacks(downstream_watermark_manager_);
+    codec_filter_->shipHeadersIfPaused(*parent_.downstreamHeaders());
+    return;
+  }
+
+  calling_encode_headers_ = true;
+  auto* headers = parent_.downstreamHeaders();
+  const auto* route_entry = parent_.route()->routeEntry();
+  if (route_entry->autoHostRewrite() && !host->hostname().empty()) {
+    Http::Utility::updateAuthority(*parent_.downstreamHeaders(), host->hostname(),
+                                   route_entry->appendXfh());
+  }
+
+  if (span_ != nullptr) {
+    span_->injectContext(*parent_.downstreamHeaders(), host);
+  } else {
+    // No independent child span for current upstream request then inject the parent span's tracing
+    // context into the request headers.
+    // The injectContext() of the parent span may be called repeatedly when the request is retried.
+    parent_.callbacks()->activeSpan().injectContext(*parent_.downstreamHeaders(), host);
+  }
+
+  upstreamTiming().onFirstUpstreamTxByteSent(parent_.callbacks()->dispatcher().timeSource());
+
+  // Make sure that when we are forwarding CONNECT payload we do not do so until
+  // the upstream has accepted the CONNECT request.
+  if (protocol.has_value() &&
+      headers->getMethodValue() == Http::Headers::get().MethodValues.Connect) {
+    paused_for_connect_ = true;
+  }
+
+  const Http::Status status =
+      upstream_->encodeHeaders(*parent_.downstreamHeaders(), shouldSendEndStream());
+  calling_encode_headers_ = false;
+
+  if (!status.ok()) {
+    // It is possible that encodeHeaders() fails. This can happen if filters or other extensions
+    // erroneously remove required headers.
+    stream_info_.setResponseFlag(StreamInfo::ResponseFlag::DownstreamProtocolError);
+    const std::string details =
+        absl::StrCat(StreamInfo::ResponseCodeDetails::get().FilterRemovedRequiredRequestHeaders,
+                     "{", StringUtil::replaceAllEmptySpace(status.message()), "}");
+    parent_.callbacks()->sendLocalReply(Http::Code::ServiceUnavailable, status.message(), nullptr,
+                                        absl::nullopt, details);
+    return;
+  }
+
+  if (!paused_for_connect_) {
+    encodeBodyAndTrailers();
+  }
 }
 
-UpstreamToDownstream& UpstreamRequest::upstreamToDownstream() { return codec_filter_->bridge_; }
+void UpstreamRequest::encodeBodyAndTrailers() {
+  if (allow_upstream_filters_) {
+    codec_filter_->shipHeadersIfPaused(*parent_.downstreamHeaders());
+    return;
+  }
+  // It is possible to get reset in the middle of an encodeHeaders() call. This happens for
+  // example in the HTTP/2 codec if the frame cannot be encoded for some reason. This should never
+  // happen but it's unclear if we have covered all cases so protect against it and test for it.
+  // One specific example of a case where this happens is if we try to encode a total header size
+  // that is too big in HTTP/2 (64K currently).
+  if (deferred_reset_reason_) {
+    onResetStream(deferred_reset_reason_.value(), absl::string_view());
+  } else {
+    // Encode metadata after headers and before any other frame type.
+    if (!downstream_metadata_map_vector_.empty()) {
+      ENVOY_STREAM_LOG(debug, "Send metadata onPoolReady. {}", *parent_.callbacks(),
+                       downstream_metadata_map_vector_);
+      upstream_->encodeMetadata(downstream_metadata_map_vector_);
+      downstream_metadata_map_vector_.clear();
+    }
+
+    if (buffered_request_body_) {
+      stream_info_.addBytesSent(buffered_request_body_->length());
+      upstream_->encodeData(*buffered_request_body_, router_sent_end_stream_ && !encode_trailers_);
+    }
+
+    if (encode_trailers_) {
+      upstream_->encodeTrailers(*parent_.downstreamTrailers());
+    }
+
+    if (router_sent_end_stream_) {
+      upstreamTiming().onLastUpstreamTxByteSent(parent_.callbacks()->dispatcher().timeSource());
+    }
+  }
+}
+
+UpstreamToDownstream& UpstreamRequest::upstreamToDownstream() {
+  if (allow_upstream_filters_) {
+    return codec_filter_->bridge_;
+  } else {
+    return *this;
+  }
+}
 
 void UpstreamRequest::onStreamMaxDurationReached() {
   upstream_host_->cluster().stats().upstream_rq_max_duration_reached_.inc();
@@ -709,7 +896,9 @@ void UpstreamRequest::clearRequestEncoder() {
   // Before clearing the encoder, unsubscribe from callbacks.
   if (upstream_) {
     parent_.callbacks()->removeDownstreamWatermarkCallbacks(downstream_watermark_manager_);
-    codec_filter_->callbacks_->removeDownstreamWatermarkCallbacks(downstream_watermark_manager_);
+    if (allow_upstream_filters_) {
+      codec_filter_->callbacks_->removeDownstreamWatermarkCallbacks(downstream_watermark_manager_);
+    }
   }
   upstream_.reset();
 }
